requirements.txt
bcrypt==4.0.1
rich==13.3.2
SQLAlchemy==2.0.20
psycopg2-binary==2.9.6
python-dotenv==1.0.0
uuid
PyJWT>=2.0.0,<3.0.0
pytz>=2021.1,<2024.0
fastapi==0.95.1
uvicorn==0.21.1
